Remove obsolete documentation and simplify project structure
