+++ docs/tests.md
@@ -31,7 +31,7 @@ Checks every frontend JavaScript file with `node --check`.
 npm run test:frontend:modules
 ```
 
-Loads scripts in the same order as `ui/index.html` inside a Node VM (via the shared `tests/frontend/script_order.cjs` helper) and verifies the Session 0 refactor guardrail contracts from `docs/frontend-maintainability-tier-2-plan.md`: every active local script tag resolves on disk with no orphaned `ui/js` files and no external scripts; local scripts use blocking classic-script execution (no `async`, `defer`, `nomodule`, or unsupported `type`); the top-level `window.LlamaGui` key set matches exactly; `flagCore`, `apiClient`, `dialogs`, `manager`, `chatTemplateSelection`, `chatUi`, and `presets` expose exactly their contract keys, all callable methods; private namespaces such as `_chatInternal` are referenced only inside their owning package; each package's `*-main.js` assembler loads after its internal contributors; and flag domains load after shared options/templates and before `definitions.js`, followed by `helpers.js`. Manager internals and shared service functions must not leak as bare globals, and the compatibility API alias must reference the shared client. Negative fixtures verify rejection of altered execution attributes, a commented-out contributor, and an undefined facade method.
+Loads scripts in the same order as `ui/index.html` inside a Node VM (via the shared `tests/frontend/script_order.cjs` helper) and verifies the Session 0 refactor guardrail contracts from `docs/frontend-maintainability-tier-2-plan.md`: every active local script tag resolves on disk with no orphaned `ui/js` files and no external scripts; local scripts use blocking classic-script execution (no `async`, `defer`, `nomodule`, or unsupported `type`); the top-level `window.LlamaGui` key set matches exactly; `flagCore`, `apiClient`, `dialogs`, `manager`, `chatTemplateSelection`, `chatUi`, and `presets` expose exactly their contract keys, all callable methods; private namespaces such as `_chatInternal` and `_managerInternal` are referenced only inside their owning package; each package's `*-main.js` assembler loads after its internal contributors; and flag domains load after shared options/templates and before `definitions.js`, followed by `helpers.js`. Manager tests also load the complete package in canonical order. Manager internals and shared service functions must not leak as bare globals, and the compatibility API alias must reference the shared client. Negative fixtures verify rejection of altered execution attributes, a commented-out contributor, and an undefined facade method.
 
 ```powershell
 npm run test:flag-definitions
@@ -142,6 +142,7 @@ Fast Node tests:
 - `preset_roving_focus_unit.cjs`: the preset list focus sequence, skipping rows in collapsed groups, roving `tabindex` bookkeeping including each row's inner controls, clamped Up/Down and Home/End movement, restoring position across a re-render, and syncing the roving position when focus arrives by click or programmatic `focus()`.
 - `manager_model_cache_unit.cjs`: the shared known-model-name cache — lowercased `.gguf` names only, an empty Set for an empty models folder versus `null` for an unknown one, cache clearing on a failed refresh, stale callers adopting the winning refresh result, and the injected model-presence notification firing on both the success and failure paths.
 - `manager_model_dir_unit.cjs`: native-picker cancellation, set/status/model-refresh sequencing and races, retained launch state after partial save failures, persistent operation errors, status-based restart readiness, stale-selection clearing, shared root-state updates, command-preview rebuilding, and safe folder/error rendering.
+- `manager_lifecycle_unit.cjs`: canonical Manager-package evaluation/configuration without DOM access, requests, or timers; dependency replacement after initialization; quit cancellation and sidebar wiring; restart/readiness/cache-busted reload; failure recovery for quit/restart controls; and successful app-update handoff to GUI lifecycle while retaining the app-update status renderer. Uses only injected API fixtures and recorded timers, with production test hooks disabled.
 - `shared_services_unit.cjs`: inert service loading; API cache defaults/overrides, method/body/header/signal forwarding, literal JSON null, invalid JSON and HTTP/network failures; confirmation/prompt keyboard, backdrop and cancel behavior, safe text assignment, focus placement, empty-versus-cancel results, and listener cleanup across repeated use.
 - `manager_releases_unit.cjs`: backend selection, backend-aware release fetching, `fetchJson` cache bypass, and installed-backend summary rendering. Custom slot cases cover backend-provided labels/paths, explicit activation payloads, duplicate/polling guards, failed activation retaining the current build, and update/repair restrictions. The suite uses the public Manager facade and gated rendering/action hooks, with injected confirmations rather than private-state mutation. It also checks inert configuration, repeated initialization, one action per click, and polling cleanup on install error/unload. The browser suite switches Custom → Custom 02 → official → Custom, including a missing-tool failure and unchanged shared flags.
 - `theme_ui_unit.cjs`: theme preference storage, `data-theme` root attribute application, unknown-theme normalization, and registry-driven color-scheme hints (asserted for every entry in `THEMES`, so a new theme with the wrong `scheme` fails here). Also covers the sidebar theme menu against a DOM stub: rendering one row per registry entry, `aria-checked`/roving `tabindex`, arrow-key wrapping, Home/End, Escape returning focus to the trigger, and outside-click dismissal. Asserts every `THEMES` entry has a matching palette block in `tokens.css`, so a theme cannot be offered in the menu while rendering as the fallback. Also enforces contrast floors for every theme, which is what makes adding a theme safe rather than merely cheap:
