+++ tests/frontend/manager_model_cache_unit.cjs
@@ -10,7 +10,7 @@ const vm = require("node:vm");
 
 const ROOT = path.resolve(__dirname, "..", "..");
 const { getScriptPaths } = require("./script_order.cjs");
-const scripts = getScriptPaths().filter(src => ["js/api-client.js", "js/manager.js"].includes(src));
+const scripts = getScriptPaths().filter(src => src === "js/api-client.js" || src.startsWith("js/manager/"));
 
 function makeElement() {
     const el = {
