+++ ui/index.html
@@ -1858,7 +1858,15 @@
 <script src="/js/config-flags-ui.js?v=revamp-1"></script>
 <script src="/js/api-client.js?v=1"></script>
 <script src="/js/dialogs.js?v=1"></script>
-<script src="/js/manager.js?v=revamp-4"></script>
+<script src="/js/manager/manager-internal.js?v=1"></script>
+<script src="/js/manager/manager-status.js?v=1"></script>
+<script src="/js/manager/manager-models.js?v=1"></script>
+<script src="/js/manager/manager-lifecycle.js?v=1"></script>
+<script src="/js/manager/manager-model-dir.js?v=1"></script>
+<script src="/js/manager/manager-app-update.js?v=1"></script>
+<script src="/js/manager/manager-backends.js?v=1"></script>
+<script src="/js/manager/manager-install.js?v=1"></script>
+<script src="/js/manager/manager-main.js?v=1"></script>
 <script src="/js/presets/presets-internal.js?v=presets-split-1"></script>
 <script src="/js/presets/presets-apply.js?v=presets-split-1"></script>
 <script src="/js/presets/presets-models.js?v=presets-split-1"></script>
